put index as default export
